Fixed Home Style
Turned the Home js file into a style js file. Fixed how the background looks and the PDL fixed box was on the home -- left it on the Gaming child page. Fixed the background tile repeating problem. Fixed the link color problem. Fixed all of the funky gaps. Cleaned up the code a bit. Added a Consulting tab to the nav bar.