Finished Office and Education
Finished the office and education pages. Need preview pics. Office page mimics education page.
Got rid of js for education because code wasn't working and switched to a different concept that does not require js.
Fixed the footer on the home page so that the <p> tag does not cover it. 
Changed the class for the footer on the gaming page so that the style page did not change it because 2rpm looks better than 8rpm on the home page.